feat: processor export phases interface
Add start/finish method to processor interface
